requirements.txt
# requirements.txt
# Install all dependencies with: pip install -r requirements.txt

streamlit>=1.35.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
